update material transfer

+++ golden/golden/reference.py
@@ -45,7 +45,7 @@ def submit_sales_order_2(doc, method):
 def submit_sales_order_3(doc, method):
     warehouse_detail = frappe.db.sql("""select `name` from `tabWarehouse` where is_group = '0' and type = 'Location' and parent is not null""", as_dict=1)
     for wd in warehouse_detail:
-        bins = frappe.db.sql("""select * from `tabBin` where warehouse = %s and projected_qty < 0""", wd.name, as_dict=1)
+        bins = frappe.db.sql("""select * from `tabBin` where warehouse = %s and (projected_qty + ito_qty) < 0""", wd.name, as_dict=1)
         for row in bins:
             check_ito = frappe.db.sql("""select count(*) from `tabITO` where docstatus = '0'""")[0][0]
             if flt(check_ito) == 0:
@@ -62,7 +62,7 @@ def submit_sales_order_4(doc, method):
     ito_id = frappe.db.sql("""select `name` from `tabITO` where docstatus = '0'""")[0][0]
     warehouse_detail = frappe.db.sql("""select `name` from `tabWarehouse` where is_group = '0' and type = 'Location' and parent is not null""", as_dict=1)
     for wd in warehouse_detail:
-        bins = frappe.db.sql("""select * from `tabBin` where warehouse = %s and projected_qty < 0""", wd.name, as_dict=1)
+        bins = frappe.db.sql("""select * from `tabBin` where warehouse = %s and (projected_qty + ito_qty) < 0""", wd.name, as_dict=1)
         for row in bins:
             check_ito_item = frappe.db.sql("""select count(*) from `tabITO Item` where parent = %s and item_code = %s and location = %s""", (ito_id, row.item_code, row.warehouse))[0][0]
             if flt(check_ito_item) == 0:
@@ -74,15 +74,16 @@ def submit_sales_order_4(doc, method):
         			"parenttype": "ITO",
                 	"item_code": row.item_code,
                 	"item_name": item_name,
-                	"qty_need": flt(row.projected_qty) * -1,
+                	"qty_need": (flt(row.projected_qty) * -1) - flt(row.ito_qty),
                     "stock_uom": row.stock_uom,
+                    "qty":0,
                     "location": row.warehouse,
                     "bin": row.name
                 })
                 ito_item.insert()
             else:
                 ito_item = frappe.get_doc("ITO Item", {"parent": ito_id, "item_code": row.item_code, "location": row.warehouse})
-                ito_item.qty_need = flt(row.projected_qty) * -1
+                ito_item.qty_need = (flt(row.projected_qty) * -1) - flt(row.ito_qty)
                 ito_item.save()
 
 def cancel_sales_order(doc, method):
@@ -93,17 +94,33 @@ def cancel_sales_order(doc, method):
     	cancel_picking.delete()
 
 def cancel_sales_order_2(doc, method):
-    ito_id = frappe.db.sql("""select `name` from `tabITO` where docstatus = '0'""")[0][0]
-    warehouse_detail = frappe.db.sql("""select `name` from `tabWarehouse` where is_group = '0' and type = 'Location' and parent is not null""", as_dict=1)
-    for wd in warehouse_detail:
-        bins = frappe.db.sql("""select * from `tabBin` where warehouse = %s""", wd.name, as_dict=1)
-        for row in bins:
-            check_ito_item = frappe.db.sql("""select count(*) from `tabITO Item` where parent = %s and item_code = %s and location = %s""", (ito_id, row.item_code, row.warehouse))[0][0]
-            if flt(check_ito_item) == 1:
-                if flt(row.projected_qty) >= 0:
-                    ito_item = frappe.get_doc("ITO Item", {"parent": ito_id, "item_code": row.item_code, "location": row.warehouse})
-                    ito_item.delete()
-                else:
-                    ito_item = frappe.get_doc("ITO Item", {"parent": ito_id, "item_code": row.item_code, "location": row.warehouse})
-                    ito_item.qty_need = flt(row.projected_qty) * -1
-                    ito_item.save()
+    count_ito = frappe.db.sql("""select count(*) from `tabITO` where docstatus = '0'""")[0][0]
+    if flt(count_ito) == 1:
+        ito_id = frappe.db.sql("""select `name` from `tabITO` where docstatus = '0'""")[0][0]
+        warehouse_detail = frappe.db.sql("""select `name` from `tabWarehouse` where is_group = '0' and type = 'Location' and parent is not null""", as_dict=1)
+        for wd in warehouse_detail:
+            bins = frappe.db.sql("""select * from `tabBin` where warehouse = %s""", wd.name, as_dict=1)
+            for row in bins:
+                check_ito_item = frappe.db.sql("""select count(*) from `tabITO Item` where parent = %s and item_code = %s and location = %s""", (ito_id, row.item_code, row.warehouse))[0][0]
+                if flt(check_ito_item) == 1:
+                    if flt(row.projected_qty) >= 0:
+                        ito_item = frappe.get_doc("ITO Item", {"parent": ito_id, "item_code": row.item_code, "location": row.warehouse})
+                        ito_item.delete()
+                    else:
+                        ito_item = frappe.get_doc("ITO Item", {"parent": ito_id, "item_code": row.item_code, "location": row.warehouse})
+                        ito_item.qty_need = flt(row.projected_qty) * -1
+                        ito_item.save()
+
+def cancel_sales_order_3(doc, method):
+    count_ito = frappe.db.sql("""select count(*) from `tabITO` where docstatus = '0'""")[0][0]
+    if flt(count_ito) == 1:
+        ito_id = frappe.db.sql("""select `name` from `tabITO` where docstatus = '0'""")[0][0]
+        count_ito_item = frappe.db.sql("""select count(*) from `tabITO Item` where `name` = %s""", ito_id)[0][0]
+        if flt(count_ito_item) == 0:
+            delete_ito = frappe.get_doc("ITO", ito_id)
+            delete_ito.delete()
+
+def submit_stock_entry(doc, method):
+    if doc.ito:
+        frappe.db.sql("""update `tabBin` set ito = null, ito_qty = 0 where ito = %s""", doc.ito)
+        frappe.db.sql("""update `tabITO` set status = 'Completed' where `name` = %s""", doc.ito)
